@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -8,8 +8,8 @@ using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using Microsoft.EntityFrameworkCore;
 using Steeltoe.CloudFoundry.Connector;
 using Steeltoe.CloudFoundry.Connector.MySql.EFCore;
@@ -34,52 +34,54 @@ namespace bootcamp_webapi
         {
             var isMySqlBound = Configuration.GetServiceInfos<MySqlServiceInfo>().Any();
             services.AddDbContext<ProductContext>(options =>
+            {
+                if (isMySqlBound)
+                    options.UseMySql(Configuration);
+                else
+                    options.UseSqlite("DataSource=:memory:");
+
+            }, isMySqlBound ? ServiceLifetime.Scoped : ServiceLifetime.Singleton);
+
+            var apiSettings = Configuration
+            .GetSection("api")
+            .Get<ApiSettings>();
+
+            services.AddSwaggerDocument(config => 
+            {
+                config.PostProcess = document =>
                 {
-                    if (isMySqlBound)
-                        options.UseMySql(Configuration);
-                    else
-                        options.UseSqlite("DataSource=:memory:");
+                    document.Info.Version = apiSettings?.Version;
+                    document.Info.Title = apiSettings?.Title;
+                    document.Info.Description = "A simple ASP.NET Core web API";
+                    document.Schemes.Clear();
+                    document.Schemes.Add(NSwag.OpenApiSchema.Https);
+                };
+            });
 
-                }, isMySqlBound ? ServiceLifetime.Scoped : ServiceLifetime.Singleton);
-
-            services.AddSwagger();
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddControllers();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
-            else
-            {
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
+
+            app.UseOpenApi();
+            app.UseSwaggerUi3(settings => settings.Path = "");
 
             app.UseHttpsRedirection();
 
-            var apiSettings = Configuration
-                .GetSection("api")
-                .Get<ApiSettings>();
+            app.UseRouting();
 
-            app.UseSwaggerUi3WithApiExplorer(settings =>
-                {
-                    settings.GeneratorSettings.DefaultPropertyNameHandling =
-                        PropertyNameHandling.CamelCase;
-                    settings.PostProcess = document =>
-                    {
-                        document.Info.Version = apiSettings?.Version;
-                        document.Info.Title = apiSettings?.Title;
-                        document.Info.Description = "A simple ASP.NET Core web API";
-                        document.Schemes.Add(NSwag.SwaggerSchema.Https);
-                    };
-                    settings.SwaggerUiRoute = "";
-                });
+            app.UseAuthorization();
 
-            app.UseMvc();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
