@@ -8,25 +8,29 @@ using bootcamp_webapi;
 namespace bootcamp_webapi.Migrations
 {
     [DbContext(typeof(ProductContext))]
-    [Migration("20190808075732_InitialCreation")]
+    [Migration("20200708082457_InitialCreation")]
     partial class InitialCreation
     {
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.2.6-servicing-10079");
+                .HasAnnotation("ProductVersion", "3.1.4");
 
             modelBuilder.Entity("bootcamp_webapi.Product", b =>
                 {
                     b.Property<long>("Id")
-                        .ValueGeneratedOnAdd();
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("INTEGER");
 
-                    b.Property<string>("Category");
+                    b.Property<string>("Category")
+                        .HasColumnType("TEXT");
 
-                    b.Property<int>("Inventory");
+                    b.Property<int>("Inventory")
+                        .HasColumnType("INTEGER");
 
-                    b.Property<string>("Name");
+                    b.Property<string>("Name")
+                        .HasColumnType("TEXT");
 
                     b.HasKey("Id");
 
