@@ -1,12 +1,12 @@
 using System.Data.Common;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 public static class EnsureMigrations
 {
     static DbConnection _connection;
-    public static void EnsureMigrationOfContext<T>(this IWebHost webHost) where T : DbContext
+    public static void EnsureMigrationOfContext<T>(this IHost webHost) where T : DbContext
     {
         using (var scope = webHost.Services.CreateScope())
         using (var serviceScope = scope.ServiceProvider.GetRequiredService<IServiceScopeFactory>().CreateScope())
