@@ -1,14 +1,13 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Steeltoe.Extensions.Configuration.ConfigServer;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace bootcamp_webapi
 {
@@ -16,16 +15,18 @@ namespace bootcamp_webapi
     {
         public static void Main(string[] args)
         {
-            IWebHost webHost = CreateWebHostBuilder(args).Build();
-            webHost.EnsureMigrationOfContext<ProductContext>();
-            webHost.Run();
+            IHost host = CreateHostBuilder(args).Build();
+            host.EnsureMigrationOfContext<ProductContext>();
+            host.Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseCloudFoundryHosting()
-                .AddConfigServer(GetLoggerFactory())
-                .UseStartup<Startup>();
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.AddConfigServer(GetLoggerFactory());
+                    webBuilder.UseStartup<Startup>();
+                });
 
         public static ILoggerFactory GetLoggerFactory()
         {
